VotingViewModel 구현 및 적용 #64
